requirements.txt
python-telegram-bot==20.0
pymongo==4.3.3
python-dotenv==1.0.0
loguru==0.6.0
